feat : 다중 파일 업로드 시, 파일명 중복 방지 추가

+++ src/main/java/com/dalrun/util/MultiFileUtil.java
@@ -3,6 +3,7 @@ package com.dalrun.util;
 import java.io.BufferedOutputStream;
 import java.io.File;
 import java.io.FileOutputStream;
+import java.util.Date;
 import java.util.List;
 
 import org.springframework.web.multipart.MultipartFile;
@@ -29,15 +30,32 @@ public class MultiFileUtil {
 		    
 			File file = new File(filepath);
 			
-			try {
-				BufferedOutputStream bos = new BufferedOutputStream(new FileOutputStream(file));
-				bos.write(files.get(i).getBytes());
-				bos.close();
-			} catch (Exception e) {
-				System.out.println(e);
-				return false;
-			} 
+			if(!file.exists()) {
+				try {
+					BufferedOutputStream bos = new BufferedOutputStream(new FileOutputStream(file));
+					bos.write(files.get(i).getBytes());
+					bos.close();
+				} catch (Exception e) {
+					System.out.println(e);
+					return false;
+				} 
+			}
 		}
 		return true;
+	}
+	
+	public static String getMultiFileName(String filename, int i) {
+		String newfilename = "";
+        String fpost = "";
+            	
+    	if (filename.indexOf('.') >= 0) {    // 확장자명이 있음
+    		fpost = filename.substring(filename.indexOf('.')); 
+    		newfilename = new Date().getTime() + i + fpost;    
+    	}
+    	else {     // 확장자명이 없음
+    		newfilename = new Date().getTime() + i + ".back"; 
+    	}
+        
+        return newfilename;
 	}
 }
